+++ src/main/java/com/todoay/api/domain/auth/service/AuthServiceImpl.java
@@ -4,10 +4,12 @@ package com.todoay.api.domain.auth.service;
 import com.todoay.api.domain.auth.dto.AuthSaveDto;
 import com.todoay.api.domain.auth.dto.AuthUpdatePasswordReqeustDto;
 import com.todoay.api.domain.auth.dto.LoginRequestDto;
+import com.todoay.api.domain.auth.dto.LoginResponseDto;
 import com.todoay.api.domain.auth.entity.Auth;
 import com.todoay.api.domain.auth.exception.EmailDuplicateException;
 import com.todoay.api.domain.auth.repository.AuthRepository;
 import com.todoay.api.domain.profile.exception.EmailNotFoundException;
+import com.todoay.api.global.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -23,6 +25,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class AuthServiceImpl implements AuthService {
 
     private final AuthRepository authRepository;
+    private final JwtProvider jwtProvider;
 
     // spring security 필수 메소드
     @Override
@@ -47,7 +50,10 @@ public class AuthServiceImpl implements AuthService {
 
     @Transactional
     @Override
-    public void updateAuthPassword(String email, AuthUpdatePasswordReqeustDto dto) {
+    public void updateAuthPassword(AuthUpdatePasswordReqeustDto dto) {
+
+
+        String email = jwtProvider.getLoginId();
         Auth auth = authRepository.findByEmail(email)
                 .orElseThrow(EmailNotFoundException::new);
 
@@ -61,8 +67,8 @@ public class AuthServiceImpl implements AuthService {
 
     @Transactional
     @Override
-    public void deleteAuth(String email) {
-
+    public void deleteAuth() {
+        String email = jwtProvider.getLoginId();
         Auth auth = authRepository.findByEmail(email)
                 .orElseThrow(EmailNotFoundException::new);
         auth.deleteAuth();
@@ -71,12 +77,15 @@ public class AuthServiceImpl implements AuthService {
     }
 
     @Override
-    public void login(LoginRequestDto loginRequestDto) {
+    public LoginResponseDto login(LoginRequestDto loginRequestDto) {
         Auth auth = (Auth) loadUserByUsername(loginRequestDto.getEmail());
         BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
         if (!encoder.matches(loginRequestDto.getPassword(), auth.getPassword())) {
             throw new IllegalArgumentException();  // 나중에 custom exception 추가
         }
+        String accessToken = jwtProvider.createAccessToken(loginRequestDto.getEmail());
+        String refreshToken = jwtProvider.createRefreshToken(loginRequestDto.getEmail());
+        return new LoginResponseDto(accessToken,refreshToken);
     }
 
     @Override
