+++ src/main/java/com/todoay/api/global/jwt/JwtProvider.java
@@ -37,7 +37,6 @@ public class JwtProvider {
     private final long REFRESH_TOKEN_EXPIRATION = 1000 * 60 * 60 * 24 * 30;
 
     private final Logger LOGGER = LoggerFactory.getLogger(JwtProvider.class);
-    private final UserDetailsService userDetailsService;
 
     @PostConstruct // init() 메소드
     protected void init() {  // secretKey를 Base64형식으로 인코딩함. 인코딩 전후 확인 로깅
@@ -77,16 +76,6 @@ public class JwtProvider {
     public String createRefreshToken(String email) {
         // refreshToken 저장해줘야한다.
         return createToken(REFRESH_TOKEN_EXPIRATION, email);
-    }
-
-
-    // 필터에서 인증이 성공했을 때 SecurityContextHolder에 저장할 Authentication을 생성하는 역할
-    // 이걸 구현하는 편한 방법은 UsernamePasswordAuthenticationToken을 사용하는 것
-    public Authentication getAuthentication(String token) {
-        LOGGER.info("[getAuthentication] 토큰 인증 정보 조회 시작");
-        UserDetails userDetails = userDetailsService.loadUserByUsername(this.getUserEmail(token));
-        LOGGER.info("[getAuthentication] 토큰 인증 정보 조회 완료, UserDetails UserName : {}", userDetails.getUsername());
-        return new UsernamePasswordAuthenticationToken(userDetails, "", userDetails.getAuthorities());
     }
 
     public String getUserEmail(String token) {
